+++ packages/hardhat-plugin-fireblocks/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @appliedblockchain/silentdatarollup-hardhat-plugin-fireblocks
 
+## 1.0.7
+
+### Patch Changes
+
+- Signature requirement on eth_getTransactionCount removed
+- Private events contract added to examples
+- Tests added
+- Updated dependencies
+  - @appliedblockchain/silentdatarollup-core@1.0.7
+
 ## 1.0.6
 
 ### Patch Changes